Introduce the libcall calling convention.
Exposed to C with `__attribute__((cheri_libcall))`, this allows calling
stateless (no access to read-write globals) functions that can be shared
between compartments.  The argument convention is the same as the normal
(within a compartment) calling convention but these functions are called
via a sealed capability found in the import table.

Calling a libcall from the library pseudo-compartment is a simple direct
call.

Library-compartment calls are differentiated from calls that go via the
compartment switches by the low bit being set to 1 in the import-table
entry.

Also add some warnings if you define a libcall from a compartment or
define a compartment-exported function from a different compartment.

This commit also fixes a couple of the test cases that were missed
previously.